[Kernels][GPU] Add stream arg to event record (#26541)
CUDA events are recorded on a stream. There is no version of
cuEventRecord that does not take a stream, so I'm not sure what was
being passed in before. Maybe the default stream was used but the
kernels being timed were not launched on the default stream.

Co-authored-by: Ian Tramble <ian@modular.com>
GPU_TEST_ORIG_REV_ID: ec7afc20d6a203733bd6f6c29dabca211ca5d6bb