[Kernels][GPU] Try using elementwise in mandelbrot demo (#26729)
This uses a new variant of elementwise that doesn't take an out_chain.
The function applied elementwise is shared between the CPU and GPU.

One issue is that using elementwise for the CPU version regresses perf
from 65ms->270ms (GPU time is the same). I think this is probably
because elementwise gives each CPU thread a chunk of rows, but the work
done by each row is not evenly distrubuted. We can't overpartition with
elementwise.

Adding this as a v2 option for now.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
ALGORITHM_ORIG_REV_ID: 86b83d15cf4dfe5779181760fa84cde2eb305962