Revert "[Kernels] Deprecate async_parallelize" (#27736)
Reverts [Internal link]

Causing significant regressions in DLRM because the group of gathers,
which each use a single thread, are serialized. This happens because GRT
launches all successors of a node in the same thread (bunch of gathers
have same predecessor). When the gather becomes sync, it blocks the
launch of other gathers.

@mbs-modular and I have a plan to reland this with some additional
annotation on the kernel that make it async when needed.

BUFFER_ORIG_REV_ID: 64c7dbd9b0ee5c59d58fbe2edccbe81f04ef2e09