[CUDA] Some CUDA stream via KGEN::OutputChain polish (#27748)
This PR started as eliminating the CUstream from the KGEN::OutputChain
entirely, using a new Mojo Stream.get_current_stream() method which
pulls the stream from TLS, and establishing that global state in the
launch primitive on the C++ side.

After looking at that, decided with Ian we should retain the
KGEN::OutputChain (though we'll rename it to KGEN::CallContext or
something) as a way to convey target-dependent call context from the C++
runtime into the Mojo kernel. Currently, that is just the CUDA stream
(and even that may go away if we can push stream management into Mojo
itself). For CPU it we may wish to put the 'max threads to use' hint in
there also.

So this PR is now just some trivial polish of the existing stream
management.

ALGORITHM_ORIG_REV_ID: 744525fed83da664fcc0aa64c2a1fbc4a34296ed