[Kernels] Make pack_b take static shape (#27359)
This allows the pack_b function take the shape of the a, b, and c
matrices. This allows one to have a different packing strategy based on
the shape of the matrices --- enabling extra optimizations. This is a
prereq for #26684

LINALG_ORIG_REV_ID: 1ef2726616af2e3c71a0074ebf4699ee88ee4338