[Kernels] Move out_chain.mark_error to lib boundary, pt 3 (#28049)
Removes instances of out_chain.mark_error in the exported kernels, which
are only used for legacy tests and benchmarking (not the engine). After
this all instances of out_chain.mark_error are isolated in MOGG.mojo.
For the exported functions, we trap instead of marking an error on the
out_chain until an error handling ABI is available.

Part of removing the out_chain
[Internal link]

LINALG_ORIG_REV_ID: 05f7f6b607c75ef13e2c667b4947683e61782a1c