[Kernels] Use vectorize inplace of a handrolled operation in packA, NFC (#28208)
In the packA function in the i8mm matmul, we can use vectorize inplace
of a handrolled version. This simplifies the code a bit. We can do the
same thing in the inner loop as well which will be examined in a
subsequent PR.

LINALG_ORIG_REV_ID: 3020f1d2b0f1cb86c87b2c7ce229ab5517ee84fa