[mojo-lang] Remove `__takeinit__` special form. (#28251)
The transfer operator delegates to a lifetime end or take init
depending on what it is, and CheckLifetimes tried to use
`__moveinit__` or `__takeinit__` depending on what was available.
This produced a bunch of problems, including an inconsistent
programming model and murky behavior pointed out in #26060

Given that there are ~no users of this, we decided to remove it
and reconsider this down the road.  This fixes #26060

GPU_ORIG_REV_ID: 33c74ed92ac5b91cd563c41e2b3141076b1084cc