[mojo-stdlib] Package all standard library packages together in a single 'stdlib' package (#28779)
The current state of the standard library is a bit of a spiderweb, where packages
can cyclically depend on each other and there's no clear layering. This makes
separate packaging near impossible because we can't cleanly generate build
artifacts for the different packages. This issue spawned the original introduction
of the `parsing-stdlib` flag, which disables considering binary packages for import,
but has also provided a very easy to use escape hatch for packaging issues. This
isn't possible for users though (we don't ship a source standard library package, or
any other package for that matter), so this escape hatch has created lots of time
bombs when it comes to actually shipping these features (we need to retroactively
try to unpack what the issues are and where they originate, often on tight schedules).
Outside of those issues, the current use of `parsing-stdlib` has also blocked the ability
to fully develop some features (like global variables) which wants packages to have proper
binary dependencies (so that things have a single home in the binary).

This PR aims to help remove the `parsing-stdlib` hack by introducing a single `stdlib`
package that contains all of the standard library packages. The `stdlib` package is
never really imported directly, some checks in the compiler allow users to import things
like `sys` just as before (i.e., no changes to user code should be necessary). This will allow
for us to more wholisitically unravel the layering of the standard library without continuing to
build on a stack of cards.

GPU_TEST_ORIG_REV_ID: 328a3545a05492b7de9de407f6802b8664b0a200