[Kernels] Add a kernels_utils package (#29999)
This adds a new kernels_utils package. This package will be used to
house useful functions that are used by kernel developers. To start
with,
this PR adds a `compile_code` function that allows one to compile
and print a function to assembly from within mojo code. This allows
the developers to examine the assembly of a specific function and
instantiation within the code.

COMPILE_TESTS_ORIG_REV_ID: 861b57416db266c430372b658440e62d99b1f681