[Kernels][GPU] Allow one to pass in literals to GPU functions (#29961)
A massive footgun when launching cuda functions is passing in literals.
This changes the code to capture those literals and pass them as Int or
Float32 types to the function.

Note that this does not remove the entire footgun, since calling the
function with an IntLiteral while the function takes a Float32 is not
going to work because we cannot reflect on the func_type. E.g.

```
    func(
        stream,
        (length // block_dim),
        (block_dim),
        out_device,
        in_device,
        33,
        length,
    )
```

in the `test/gpu/add_constant.mojo` is going to produce the wrong
result.

Closes #24598

GPU_TEST_ORIG_REV_ID: 9fbe0c1be3e0add3ccc07df298259aef7cafa0b4