[Runtime][Performance] tcmalloc special build (#28166)
Most of the testing is done and it looks like we have UT's and several
build flows passing with TCMalloc. TCMalloc integration closely follows
other third-party integrations like stock-tf, stock-pt etc. The flow is
as follows.
1) There is a master CMAKE Option `MODULAR_USE_TCMALLOC=<ON/OFF>` which
controls whether we link with tcmalloc in the stack.
2) There is `build-tcmalloc` analogous to build-external-tensorflow
which downloads and builds tcmalloc on Debian machines(ARM and X86)

The flow is as follows
1) source utils/start-modular.sh
2) build-tcmalloc (builds and installs tcmalloc in
.derived/third-party/gperftools)
3) cmake-modular ; build

TCMalloc is not linked when
1) MODULAR_USE_TCMALLOC=OFF or
2) TCMALLOC is not found in .derived/third-party/gperftools or
3) OSTYPE is not linux-gnu

This PR enables tcmalloc by default. There is an Engine API option to
enable disable through
`M_setAllocatorType(config, kCaching/kSystem)` -> C API
The same is extended to `python` and `mojo` api too.

Currently we are enabling the compile option but not the runtime option
To use tcmalloc please set , --allocator=tcmalloc or use the kCaching option
of the C API.
---------

Co-authored-by: Tyler Kenney <tyler@modular.com>
Co-authored-by: Tyler J. Kenney <tjk213@gmail.com>
MOJO_MAX_ORIG_REV_ID: 82c8eea02f519ff863d6a827b79f8fdb28bc0910