[Stdlib] Reorder `unroll` parameters so that func is first arg (#30874)
This reorders the `unroll` parameters so that the input function is the
first argument. This algins the function with `vectorize` and would
allow us to support any arity unroll (instead of the up to 3 loop levels
that we currently support).

MOGG_ORIG_REV_ID: 5478f50e68b8663065799078bf573c3c495174de