[MOGG] Only run brute force deducer if params need to be deduced (#31234)
Our parameter deduction is built on several layers of different methods
to deduce the parameters. The most brittle of these is the last layer of
defence which looks at the KGEN type and the type on the LIT sig and
tries to line them up. It has very few actual uses and we will probably
remove it in the near term.

Until then we should at least only run this if we actually have
something which still needs to be deduced.

MOGG_ORIG_REV_ID: 150ebecd1f13d47d6308e8de94f228112adc3e9e