[Kernels][Stdlib] Remove most uses of llcl.Runtime (#31530)
The llcl.Runtime is an internal detail and has a bunch of sharp edges
that users (including us) should not rely and use directly. Most of the
users were not even using the runtime, since it was a relic of when the
runtime was not always created and/or when parallelize took in a
runtime.

NN_ORIG_REV_ID: 998dbccddb678d5d3217feb372a2aa6561f25506