[mojo-stdlib] Move GPU tests using NN to `test/nn/gpu` (#31599)
`NN` is a kernels-related thing, not a stdlib thing even though the GPU package
is in the stdlib.  So, move these GPU tests that use `NN` to
`Kernels/test/nn/gpu`.

This was the last use of `NN` in the `mojo-stdlib` test suite.  So, go ahead and
remove the CMake dependency of `check-mojo-stdlib` on `MojoNN`.

Fixes [Internal link]

NN_ORIG_REV_ID: 4d5f60183be767b9e4164f67fe71e021d14b7557