[SDK] [API] [GenericML] Enable the use of an existing runtime (#31595)
Currently, the SDK will always attempt to unilaterally create a new runtime
instance. This requires the use of the `LLCL::createNestedRuntime` call, even
if no runtime currently exists.

Since changes for #31548 are forcing explicit lifecycle control of the runtime,
this needs to change. This commit adds an explicit API toggle that can be used
to refer to an *existing* runtime. Subsequenc changes will enforce that there a
top-level runtime in Mojo programs (either injected from `main`, or by the
engine when it is running), so this flag will be universally set by the Mojo
SDK.

This also has the advantage that the runtime configuration doesn't need to go
through the C API for Mojo either, you just need to ensure that the session
exists within scope of a runtime that you want to use for that session.

```
with Runtime(4, "-") as rt:
   with InferenceSession() as sess:
       ... # sess will have paralleism of four, refering to the
           # the rt runtime above. It is also the same runtime
           # that will execute Mojo code in this context!
```

MOJO_MAX_ORIG_REV_ID: 0ea60ee30dd92896fc18cc42566e0d86a507bcba