[MIN] Mark the mlir module private, rename capi to c. (#32354)
A number of `mlir.*` symbols are leaking into the public docs.
Technically, we should alias them under an opaque identifier, but that
seems quite invasive. Instead, mark the entire module private (something
we need to do anyway), and hope that the appearabce of a `_mlir.Foo` in
the docs is sufficient to indicate this is an undocummented symol with
no public contract.

Note: we need to scrub the docs to remove references to these as well
(and unstead refer to some "opqaue IR" types - that is for a future PR).
MOJO_MAX_ORIG_REV_ID: 1ddef850e083c5ee4013f186db5fe496ffc857f1