[Kernels] Allows matmul to dynamically pass M for benchmarking mixed kernels (#32840)
Reverts #32669 which is a revert of #31966. PR #31966 had to be reverted
due to a bug in the graph compiler #32933.

Creates a new matmul_M function which allows passing the value of M in
MxNxK matmul. This allows using mixed kernels dynamically for
benchmarking. The old matmul function only uses M if it's static.

Fixes #31965

LINALG_ORIG_REV_ID: bbdc8ecea8788f579540576694dc711b136f97f2