[MOGG] Lazy loading of mojo kernel libraries into the graph compiler to improve the compilation speed (#32613)
Implemented the lazy loading of the mojo kernel libraries to improve the
compilation speed:
- We have two modes: eager and lazy loading. Clients could pick. Eager loading has fast bulk loading and lazy loading saves by on-demand materialization.
- For a typical `mlir` FileCheck test file in our unittest, if we
compile to `mo-to-mogg` the speedup is ~3x and for `mo-to-mgp` the
saving is ~30%. This is measured when cache is enabled.
- This implementation leverages the `mlir::BytecodeReader` APIs to
materialize OPs on-demand. Specifically, it only collects the top-level
kernel generator's metadata without touching the KGEN ops inside at the
initialization. In the meantime, we have some bookkeepings to store the
state about what OPs are fully materialized and their associated
readers. The `getMojoBindings` API forces the full materialization of
the libraries. Otherwise, each OP is loaded when we explicitly require
it (e.g., via `getRegisteredMojoKernelFor` or other APIs).
- Re-worked some of the APIs to make sure all passes that use
`MojoLibraryAnalysis` can enjoy the lazy loading. For example, putting
the original `mojobindings` object into a getter method to explicitly
trigger full materialization, API to iterate over the raw mojo kernels
(without full materialization), refining the semantic of
`markAllFunctionsAsNotExported`, and the API to query the raw kernel
operator (via `lookup`).
- We made 'MojoLibraryAnalysis' non-copyable and non-movable and changed
the factory method to be constructor-based with `.init()` to cope with
the internal implementation.
- This PR relates to
[Internal link]

MOGG_ORIG_REV_ID: eba0c5fe163d873e7cc736f72c43e8c5203d92a5