[mojo-stdlib][All] Rename `trap` to `abort` (#33331)
As part of [Internal link] in cleaning
up the `debug` module, rename `trap` to `abort`. In a future PR, we'll
move these two functions to the `os` module, but the rename is
intentionally done atomically separate from the file move.

After that, we can decide what to do the `debug` module and the LLDB
visualization decorator for the final fate of the `debug` module.

KERNEL_BENCHMARKS_ORIG_REV_ID: 52e4bdc321edc037c0e8b3e263046c0826aa5bba