[Stdlib] Remove GPU print function and add support for sep/end (#33554)
This removes the GPU print function (which will be moved to a different
file in a subsequent PR). It removes the specialized unary overloads of
print, because they were only really useful for GPUs.

With these changes, we can actually have #32569 to work, so just do that
and support sep/end in the variadic print overload.

Closes #12811

QUANTIZATION_ORIG_REV_ID: 8c911f6dbb3007f931e796a549a8e33ba27cf453