[Kernels][GPU] Fix perf bug in specialized concat kernel (#34858)
[Internal link] added a specialized
concat kernel for when the axis is the innermost dim and the inputs all
have an inner dim of 1. The kernel iterates over all inputs so you only
need a thread for each coord in the input shape, not each coord in the
output shape. This speeds up the kernel by a factor of num_inputs.

This specialization should be removed soon once the general concat perf
is improved.

GPU_TEST_ORIG_REV_ID: a95b9b97d76cb2cc0a91740079e6fffc92ba0811