[LLCL] [mojo] Fix taskgrouptasklist to return a pointer to task (#35269)
In #35114 the implementation of `__getitem__` move-returns the task
which can create many unexpected errors for users. For instance, calling
`__getitem__` twice will fail since the task was moved in the first
call, similarly the destructor will be called twice.
This PR fixes the implementation to return a pointer to the task group.
The user needs to be careful to not dereference the pointer once the
taskgrouptasklist is out of scope, but that is a rather straightforward
requirement.
The task group task list can be used as follows:
```
with Runtime(4) as rt:
        var tg = TaskGroup(rt)
        var tasks =TaskGroupTaskList[T]10)
        for t in range(1,10):
            tasks.add(tg.create_task[T](func()))
        tg.wait()
        for t in range(tasks.size):
            res.append(tasks[t].load().get())
```

RUNTIME_ORIG_REV_ID: cab6417e07bda03ac631f257a47169d6fd125bcd